requirements.txt
streamlit
boto3
pandas
python-dotenv
